l2_packet_haiku: Do not specify MSG_TRUNC flag.
It doesn't have any meaning for recvfrom() in POSIX, but it
does as a Linux extension. Haiku has never implemented it,
and now returns errors when it is specified. The implementation
of this same function in l2_packet_linux does not specify MSG_TRUNC,
at least at present, so it's not needed.

Will fix https://dev.haiku-os.org/ticket/18685